requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
docopt==0.6.2
gunicorn==20.1.0
num2words==0.5.11
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
python-dotenv==0.20.0
pytz==2022.1
reportlab==3.6.12
sqlparse==0.4.2
toml==0.10.2
